+++ gazebo/gui/plot/IncrementalPlot_TEST.hh
@@ -14,26 +14,26 @@
  * limitations under the License.
  *
 */
-
-#ifndef _TEST_INTEGRATION_MODELMANIPULATION_HH_
-#define _TEST_INTEGRATION_MODELMANIPULATION_HH_
+#ifndef GAZEBO_GUI_PLOT_INCREMENTALPLOT_TEST_HH_
+#define GAZEBO_GUI_PLOT_INCREMENTALPLOT_TEST_HH_
 
 #include "gazebo/gui/QTestFixture.hh"
 
-/// \brief Integration test for manipulating models through the GUI.
-class ModelManipulationTest : public QTestFixture
+/// \brief A test class for the IncrementalPlot widget.
+class IncrementalPlot_TEST : public QTestFixture
 {
   Q_OBJECT
 
-  /// \brief Verify that model pose stops being updated while it is being
-  /// manipulated.
-  private slots: void StopProcessingPoseMsgs();
+  /// \brief Test adding and removing plot curves
+  private slots: void AddRemoveCurve();
 
-  /// \brief Test switching modes using keyboard shortcuts.
-  private slots: void Shortcuts();
+  /// \brief Test detaching and attaching curves
+  private slots: void AttachDetachCurve();
 
-  /// \brief Test manipulating models in local and global frames
-  private slots: void GlobalLocalFrames();
+  /// \brief Test adding points to curves in the plot.
+  private slots: void AddPoint();
+
+  /// \brief Test changing the curve label.
+  private slots: void SetCurveLabel();
 };
-
 #endif