+++ gazebo/gui/plot/PlotCanvas_TEST.hh
@@ -14,26 +14,23 @@
  * limitations under the License.
  *
 */
-
-#ifndef _TEST_INTEGRATION_MODELMANIPULATION_HH_
-#define _TEST_INTEGRATION_MODELMANIPULATION_HH_
+#ifndef GAZEBO_GUI_PLOT_PLOTCANVAS_TEST_HH_
+#define GAZEBO_GUI_PLOT_PLOTCANVAS_TEST_HH_
 
 #include "gazebo/gui/QTestFixture.hh"
 
-/// \brief Integration test for manipulating models through the GUI.
-class ModelManipulationTest : public QTestFixture
+/// \brief A test class for the PlotCanvas widget.
+class PlotCanvas_TEST : public QTestFixture
 {
   Q_OBJECT
 
-  /// \brief Verify that model pose stops being updated while it is being
-  /// manipulated.
-  private slots: void StopProcessingPoseMsgs();
+  /// \brief Test adding and removing plots
+  private slots: void AddRemovePlot();
 
-  /// \brief Test switching modes using keyboard shortcuts.
-  private slots: void Shortcuts();
+  /// \brief Test adding and removing variables
+  private slots: void AddRemoveVariable();
 
-  /// \brief Test manipulating models in local and global frames
-  private slots: void GlobalLocalFrames();
+  /// \brief Test changing variable label
+  private slots: void VariableLabel();
 };
-
 #endif