+++ gazebo/gui/plot/VariablePillContainer_TEST.hh
@@ -14,26 +14,20 @@
  * limitations under the License.
  *
 */
-
-#ifndef _TEST_INTEGRATION_MODELMANIPULATION_HH_
-#define _TEST_INTEGRATION_MODELMANIPULATION_HH_
+#ifndef GAZEBO_GUI_PLOT_VARIABLEPILLCONTAINER_TEST_HH_
+#define GAZEBO_GUI_PLOT_VARIABLEPILLCONTAINER_TEST_HH_
 
 #include "gazebo/gui/QTestFixture.hh"
 
-/// \brief Integration test for manipulating models through the GUI.
-class ModelManipulationTest : public QTestFixture
+/// \brief A test class for the VariablePillContainer widget.
+class VariablePillContainer_TEST : public QTestFixture
 {
   Q_OBJECT
 
-  /// \brief Verify that model pose stops being updated while it is being
-  /// manipulated.
-  private slots: void StopProcessingPoseMsgs();
+  /// \brief Test adding and removing variable pills
+  private slots: void AddRemoveVariable();
 
-  /// \brief Test switching modes using keyboard shortcuts.
-  private slots: void Shortcuts();
-
-  /// \brief Test manipulating models in local and global frames
-  private slots: void GlobalLocalFrames();
+  /// \brief Test container capacity setting
+  private slots: void MaxSize();
 };
-
 #endif