+++ gazebo/gui/plot/VariablePill_TEST.hh
@@ -14,26 +14,23 @@
  * limitations under the License.
  *
 */
-
-#ifndef _TEST_INTEGRATION_MODELMANIPULATION_HH_
-#define _TEST_INTEGRATION_MODELMANIPULATION_HH_
+#ifndef GAZEBO_GUI_PLOT_VARIABLEPILL_TEST_HH_
+#define GAZEBO_GUI_PLOT_VARIABLEPILL_TEST_HH_
 
 #include "gazebo/gui/QTestFixture.hh"
 
-/// \brief Integration test for manipulating models through the GUI.
-class ModelManipulationTest : public QTestFixture
+/// \brief A test class for the VariablePill widget.
+class VariablePill_TEST : public QTestFixture
 {
   Q_OBJECT
 
-  /// \brief Verify that model pose stops being updated while it is being
-  /// manipulated.
-  private slots: void StopProcessingPoseMsgs();
+  /// \brief Test various setter and accessor functions
+  private slots: void Variable();
 
-  /// \brief Test switching modes using keyboard shortcuts.
-  private slots: void Shortcuts();
+  /// \brief Test creating variable pills and verify they have unique ids
+  private slots: void VariableId();
 
-  /// \brief Test manipulating models in local and global frames
-  private slots: void GlobalLocalFrames();
+  /// \brief Test creating a mulit-variable pill
+  private slots: void MultiVariable();
 };
-
 #endif