+++ Changelog.md
@@ -8,6 +8,9 @@
 1. Fix Road2d vertices and shadows
     * [Pull request #2362](https://bitbucket.org/osrf/gazebo/pull-request/2362)
 
+1. Plotting utility - plot window, canvas, variables
+    * [Pull request #2348](https://bitbucket.org/osrf/gazebo/pull-request/2348)
+
 1. Rearrange GLWidget::OnMouseMove so that the more common use cases it
    fewer if statements. Use std::thread in place of boost in OculusWindow.
    Pragma statements to prevent warnings. Prevent variable hiding in
@@ -124,6 +127,13 @@
 ## Gazebo 7
 
 ## Gazebo 7.x.x (2016-xx-xx)
+
+1. Fix Model Manipulator switching between local and global frames
+    * [Pull request 2361](https://bitbucket.org/osrf/gazebo/pull-request/2361)
+
+1. Remove contact filters with names that contain `::`
+    * [Pull request 2363](https://bitbucket.org/osrf/gazebo/pull-request/2363)
+    * [Issue 1985](https://bitbucket.org/osrf/gazebo/issues/1805)
 
 ## Gazebo 7.2.0 (2016-06-13)